Computer Player added
characters marked as team 1 and with the Strategy Script will now act at the end of the turn and attack characters on other teams.

I need to add a time delay into the CP's actions, the model cannot move fast enough for the checks that are run after it begins moving.
Addtionaly the attacking for loop should be inside of a loop tied to energy and have delays so each attack corresponds to the animation more closely.